+++ .vscode/settings.json
@@ -1,6 +1,8 @@
 {
   "editor.formatOnSave": true,
-  "mocha.requires": ["ts-node/register"],
+  "mocha.requires": [
+    "ts-node/register"
+  ],
   "mocha.files.glob": "src/**/*.test.ts",
   "mocha.subdirectory": "zedd-app",
   "mocha.env": {
@@ -9,6 +11,9 @@
   },
   "eslint.nodePath": "zedd-app/node_modules",
   "eslint.alwaysShowStatus": true,
-  "eslint.workingDirectories": ["./zedd-app", "./zedd-clarity"],
+  "eslint.workingDirectories": [
+    "./zedd-app",
+    "./zedd-platform"
+  ],
   "editor.defaultFormatter": "esbenp.prettier-vscode"
 }
