+++ zedd-app/package.json
@@ -71,7 +71,8 @@
                   "name": "main_window"
                 }
               ]
-            }
+            },
+            "loggerPort": "9001"
           }
         }
       ],
@@ -173,6 +174,6 @@
     "tslint-config-prettier": "^1.18.0",
     "typescript": "^4.5.5",
     "unzipper": "^0.10.11",
-    "zedd-clarity": "file:../zedd-clarity"
+    "zedd-platform": "file:../zedd-platform"
   }
 }
