+++ zedd-app/webpack.renderer.config.js
@@ -1,23 +1,25 @@
 const rules = require('./webpack.rules')
 
 rules.push({
-  test: /\.css$/,
-  use: [{ loader: 'style-loader' }, { loader: 'css-loader' }],
+    test: /\.css$/,
+    use: [{loader: 'style-loader'}, {loader: 'css-loader'}],
 })
 
 module.exports = {
-  module: {
-    rules,
-  },
-  target: 'electron-renderer',
-  resolve: {
-    extensions: ['.js', '.ts', '.jsx', '.tsx', '.css'],
-  },
-  externals: {
-    'selenium-webdriver': 'commonjs selenium-webdriver',
+    module: {
+        rules,
+    },
+    target: 'electron-renderer',
+    resolve: {
+        extensions: ['.js', '.ts', '.jsx', '.tsx', '.css'],
+    },
+    externals: {
+        'puppeteer': 'commonjs puppeteer',
 
-    'selenium-webdriver/chrome': 'commonjs selenium-webdriver/chrome',
+        'selenium-webdriver': 'commonjs selenium-webdriver',
 
-    'win-ca': 'commonjs win-ca',
-  },
+        'selenium-webdriver/chrome': 'commonjs selenium-webdriver/chrome',
+
+        'win-ca': 'commonjs win-ca',
+    },
 }
