+++ zedd-platform/package.json
@@ -1,9 +1,10 @@
 {
-  "name": "zedd-clarity",
+  "name": "zedd-platform",
   "version": "1.0.0",
   "description": "",
   "main": "out/index.js",
   "scripts": {
+    "test:ott-export": "ts-node --compiler-options '{\"module\":\"CommonJS\"}' test/ott-export.ts",
     "test": "echo \"Error: no test specified\" && exit 1",
     "build": "tsc",
     "watch": "tsc --watch",
@@ -31,9 +32,11 @@
     "@types/selenium-webdriver": "^4.0.18",
     "date-fns": "^2.28.0",
     "deep-equal": "^2.0.5",
+    "devtools-protocol": "0.0.1450379",
     "fast-csv": "^4.3.6",
     "jszip": "^3.7.1",
     "lodash": "^4.17.21",
+    "puppeteer": "^19.0.0",
     "selenium-webdriver": "^4.1.1",
     "sleep-promise": "^9.1.0",
     "ts-node": "^10.5.0",
